// src/app/taf-fiscal/tsi/models/tsi-error-item.ts
export interface TsiErrorItem {
  branchcode: string;
  layouterror: string;
  integrationdate: string;
  integrationtime: string;
  regkey: string;
  erpkey: string;
  errormessage: string;
  typeInvoice?: string;
  invoice?: string;
  serie?: string;
  dateInvoice?: string;
  codeParticipant?: string;
  key?: number;
}